+++ interfaces/HO-Portal/src/app/services/programs-service-api.service.ts
@@ -129,9 +129,10 @@ export class ProgramsServiceApiService {
     programId: number | string,
     phase: ProgramPhase,
   ): Promise<PaTableAttribute[]> {
+    const phaseString = phase ? phase : '';
     return this.apiService.get(
       environment.url_121_service_api,
-      `/programs/${programId}/pa-table-attributes/${phase}`,
+      `/programs/${programId}/pa-table-attributes/${phaseString}`,
     );
   }
 
@@ -604,6 +605,14 @@ export class ProgramsServiceApiService {
     message: string,
   ): Promise<any> {
     return this.updatePaStatus('reject', programId, referenceIds, message);
+  }
+
+  pause(
+    programId: number | string,
+    referenceIds: string[],
+    message: string,
+  ): Promise<any> {
+    return this.updatePaStatus('pause', programId, referenceIds, message);
   }
 
   sendMessage(